add more #pragmas to turn of some warnings, minor cleanups
git-svn-id: http://svn.ratbox.org/svnroot/ircd-ratbox/trunk@28569 b93f080a-4cfa-0310-a19a-80a68f531ef9

+++ include/channel.h
@@ -83,8 +83,7 @@ struct membership
 
 	struct Channel *chptr;
 	struct Client *client_p;
-	uint8_t flags;
-
+	int flags;
 	uint32_t ban_serial;
 };
